% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aka_check.R
\name{aka_check}
\alias{aka_check}
\title{Functions to indicate the presence of characters signalling uncertainty}
\usage{
aka_check(n)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}
}
\value{
A logical vector of the same length as n
}
\description{
Functions to indicate the presence of characters signalling uncertainty
}
\details{
aka_check checks for strings such as 又名 which may indicate that two
 names are listed in the field;
}
