% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_4corner.R
\encoding{UTF-8}
\name{ChStr2fc}
\alias{ChStr2fc}
\title{Convert Chinese strings to four corner code.}
\usage{
ChStr2fc(Chin.strs, sep = "_", parallel = FALSE, ...)
}
\arguments{
\item{Chin.strs}{The string need to be converted}

\item{sep}{Character used to seperate different characters. With current programming, should only be '_' or ''}

\item{...}{Unused}
}
\value{
four corner code of \code{Chin.str}.
}
\description{
Convert Chinese strings to four corner code.
}
\examples{
ChStr2fc(c("\\u6d77\\u4e0a\\u751f\\u660e\\u6708","\\u5929\\u6daf\\u5171\\u6b64\\u65f6"))
}
