% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pinyin.R
\encoding{UTF-8}
\name{ChStr2py}
\alias{ChStr2py}
\title{Convert Chinese strings to pinyin.}
\usage{
ChStr2py(Chin.strs, tones = TRUE, multi = TRUE, sep = "_", ...)
}
\arguments{
\item{Chin.strs}{The string need to be converted. Note, it should have a declared encoding, preferably UTF-8}

\item{tones}{Whether the output should be toned (T) or toneless.}

\item{multi}{Whether the output should list multiple pinyins for characters with multiple pronunciations}

\item{sep}{Character used to seperate different characters. Must be '' or '_' for proper operation}

\item{...}{Unused}
}
\value{
pinyin of \code{Chin.str}.
}
\description{
Convert Chinese strings to pinyin.
}
