% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chstr2rad.R
\name{Chstr2rad}
\alias{Chstr2rad}
\title{Extract radical decomposition of Chinese Characters strings}
\usage{
Chstr2rad(n, times = 1)
}
\arguments{
\item{n}{A vector of Chinese character strings}

\item{times}{Maximum number of serial decompositions to use}
}
\value{
A vector of radical decompositions of the original strings processed
 through at most \code{times} rounds of decomposition
}
\description{
Extract radical decomposition of Chinese Characters strings
}
\examples{
Chstr2rad('凨冪',1)
Chstr2rad('凨冪',2)
}
