% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_structure.R
\name{ChStr2struct}
\alias{ChStr2struct}
\title{Extract structure information from Chinese Character strings.}
\usage{
ChStr2struct(Chin.strs = "", sep = "_", parallel = FALSE,
  full = FALSE)
}
\arguments{
\item{Chin.strs}{A vector of Chinese character strings}

\item{sep}{Separating character to be placed between results for individual Hanzi}

\item{parallel}{option to run with parallel calculations}

\item{full}{Boolean option to return either major character structure only, or layered structure information based on full radical decomposition}
}
\value{
A vector of unicode ideograph structure strings representing the
 structure of the input characters
}
\description{
Extract structure information from Chinese Character strings.
}
\examples{
ChStr2struct('凨冪',full=F)
ChStr2struct('凨冪',full=T)
}
