% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wubi.R
\encoding{UTF-8}
\name{ChStr2wb}
\alias{ChStr2wb}
\title{Convert Chinese strings to wubi code (based on radicals).}
\usage{
ChStr2wb(Chin.strs, sep = "_", ...)
}
\arguments{
\item{sep}{Character used to seperate different characters, must be '_' or '' for proper functioning}

\item{...}{unused}

\item{Chin.str}{The string need to be converted}
}
\value{
wubi code of \code{Chin.str}.
}
\description{
Convert Chinese strings to wubi code (based on radicals).
}
\examples{
ChStr2wb(c("\\u6d77\\u4e0a\\u751f\\u660e\\u67081","\\u5929\\u6daf\\u5171\\u6b64\\u65f6"))
}
