% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FClib_dataset.R
\docType{data}
\name{FClib}
\alias{FClib}
\title{Four Corner database}
\format{An environment with Chinese characters as the names.}
\source{
\href{http://bbs.unispim.com/forum.php?mod=viewthread&tid=31674}{unispim}
}
\usage{
data(FClib)
}
\description{
A lookup table for the four corner code of Chinese characters
}
\examples{
data(FClib)

}
\keyword{datasets}
