% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han_check.R
\name{hancheck}
\alias{hancheck}
\title{Rough detection of Han Chinese names}
\usage{
hancheck(n)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}
}
\value{
A logical vector of the same length as n
}
\description{
Rough detection of Han Chinese names
}
\details{
Han classification is based on two characteristics. The first is
  whether each name starts with one of the 400 most popular family names
  listed for 2013 by Baidu. The second requirement is that the name is 2-4
  characters long.
}
