% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homonym.R
\name{homonym}
\alias{homonym}
\title{Support function for Pinyin with more than one pronunciation. Expects format to be "...[py1,py2]..."}
\usage{
homonym(s)
}
\arguments{
\item{s}{A string of pinyin-converted Chinese characters}
}
\value{
A vector of all possible pronunciations for the full string \code{s}.
}
\description{
Support function for Pinyin with more than one pronunciation. Expects format to be "...[py1,py2]..."
}
\examples{
homonym('[lei1,luo4]ta2')
}
