% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_dataset.R
\docType{data}
\name{idslib}
\alias{idslib}
\title{Character decomposition dataset}
\format{An object of class \code{environment} of length 20225.}
\source{
https://github.com/cjkvi/cjkvi-ids/blob/master/ids.txt
}
\usage{
idslib
}
\description{
processed with the following code:
ids = read_csv('C:/Users/phil_collender/Documents/bleh/RLpaper/RLmanuscript/ids.csv')
ids = subset(ids, Char %in% names(pylib))
ids$Decomp6 <- ids$Decomp7 <- NULL

todo = lapply(ids,function(v) grep('\\[',v))
idscleaner = function(n){
 matches = unlist(regmatches(n,gregexpr('\\[.*?\\]',n)))
 keep = grepl('G',matches)
 n[keep] = gsub('\\[.*?\\]','',n[keep])
 n[!keep] = NA
 n
}
for(i in 1:length(todo)){
 ids[[i]][todo[[i]]] = idscleaner(ids[[i]][todo[[i]]])
}
for(i in 1:length(todo)){
 ids[[i]][is.na(ids[[i]])] <- ids[[i+1]][is.na(ids[[i]])]
 ids[[i+1]][is.na(ids[[i]])] <- NA
}

structchar = c('⿰','⿱','⿲','⿳','⿴','⿵','⿶','⿷','⿸','⿹','⿺','⿻')
ids = subset(ids, substr(Decomp,1,1) %in% structchar)
nms = ids$Char
ids = lapply(1:length(nms),function(i){
 ids[i,2]
})
names(ids) = nms
idslib = list2env(ids)
save(idslib,file = 'data/idslib.rda')
}
\keyword{datasets}
