% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_freq.R
\name{name_freq}
\alias{name_freq}
\title{Functions to calculate empirical relative frequency of observed name components. Used to infer the rarity of observed name patterns for matching.}
\usage{
name_freq(n, refdata, start = 1, end = 9999, log = T)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}

\item{refdata}{Dataset from which to calculate name component frequencies; can be a precalculated lookup table or a raw vector (will add overhead to repeated computations)}

\item{start}{initial index of name substring on which to get frequencies}

\item{end}{final index of name substring on which to get frequencies}

\item{log}{logical: whether to return log relative frequency}
}
\value{
An integer vector of the same length as n
}
\description{
Functions to calculate empirical relative frequency of observed name components. Used to infer the rarity of observed name patterns for matching.
}
\details{
for each name, returns the relative frequency (or log relative frequency) of the specified substring in the reference data
}
