% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paren_check.R
\name{paren_check}
\alias{paren_check}
\title{Functions to indicate the presence of characters signalling uncertainty}
\usage{
paren_check(n)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}
}
\value{
A logical vector of the same length as n
}
\description{
Functions to indicate the presence of characters signalling uncertainty
}
\details{
paren_check detects parentheses, which are used to show possible
 character substitutions, alternate names, or comments;
}
