% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylib_dataset.R
\docType{data}
\name{pylib}
\alias{pylib}
\title{Pinyin database}
\format{An environment with Chinese characters as the names.}
\source{
\href{https://github.com/haiwen/seahub/tree/master/seahub}{GitHub haiwen}
}
\usage{
data(pylib)
}
\description{
A lookup table for pinyin of Chinese characters
}
\examples{
data(pylib)

}
\keyword{datasets}
