% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylibnew_dataset.R
\docType{data}
\name{pylibnew}
\alias{pylibnew}
\title{Pinyin database}
\format{An environment with Chinese characters as the names.}
\source{
\href{https://blog.csdn.net/chndata/article/details/41114771}{CSDN}
}
\usage{
data(pylib)
}
\description{
A lookup table for pinyin code of Chinese characters
}
\examples{
data(pylibnew)

}
\keyword{datasets}
