% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmark_check.R
\name{qmark_check}
\alias{qmark_check}
\title{Functions to indicate the presence of characters signalling uncertainty}
\usage{
qmark_check(n)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}
}
\value{
A logical vector of the same length as n
}
\description{
Functions to indicate the presence of characters signalling uncertainty
}
\details{
qmark_check detects question marks, which are used to signal the
 place of an unknown character
}
