% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad100lib_dataset.R
\docType{data}
\name{rad100lib}
\alias{rad100lib}
\title{Radical decomposition dataset after full decomposition}
\format{An object of class \code{environment} of length 20225.}
\source{
https://github.com/cjkvi/cjkvi-ids/blob/master/ids.txt
Further processed as in the ids2struct_rad script file
}
\usage{
rad100lib
}
\description{
Radical decomposition dataset after full decomposition
}
\keyword{datasets}
