% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad1lib_dataset.R
\docType{data}
\name{rad1lib}
\alias{rad1lib}
\title{Radical decomposition dataset after one round of decomposition}
\format{An object of class \code{environment} of length 20225.}
\source{
https://github.com/cjkvi/cjkvi-ids/blob/master/ids.txt
Further processed as in the ids2struct_rad script file
}
\usage{
rad1lib
}
\description{
Radical decomposition dataset after one round of decomposition
}
\keyword{datasets}
