% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_py.R
\name{revpy}
\alias{revpy}
\title{Function to parse format used for storing character matches when original
names are written partially in pinyin}
\usage{
revpy(n, all = F, py = F)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}

\item{all}{A logical indicator. If set to TRUE, all possible names are
returned. If false, only the first character match is returned.}

\item{py}{A logical indicator. If set to TRUE, the original pinyin element is
returned instead of possible characters}
}
\value{
A list of parsed names
}
\description{
Function to parse format used for storing character matches when original
names are written partially in pinyin
}
\examples{
#original name: chen1\\u6e9c
revpy('{chen1,[\\u55d4,\\u8ac3,\\u741b,\\u8cdd,\\u8b13,\\u62bb,\\u90f4,\\u778b,\\u6375,\\u7628,\\u7e1d,\\u7d9d,\\u68fd]}\\u6e9c', all=F)
revpy('{chen1,[\\u55d4,\\u8ac3,\\u741b,\\u8cdd,\\u8b13,\\u62bb,\\u90f4,\\u778b,\\u6375,\\u7628,\\u7e1d,\\u7d9d,\\u68fd]}\\u6e9c', all=T)
revpy('{chen1,[\\u55d4,\\u8ac3,\\u741b,\\u8cdd,\\u8b13,\\u62bb,\\u90f4,\\u778b,\\u6375,\\u7628,\\u7e1d,\\u7d9d,\\u68fd]}\\u6e9c', py=T)
}
