% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_func_mat.R
\name{sim_func_mat}
\alias{sim_func_mat}
\title{Compute a matrix of similarity measures between strings, accounting for ambiguous
characters. In situations where all combinations or permutations of a set
of strings are to be compared, this implementation is more efficient than calling sim_func
However, it may be less memory efficient for large inputs}
\usage{
sim_func_mat(s_1, s_2, method = "lv", q = 1, ...)
}
\arguments{
\item{s_1}{A vector or matrix of characters representing one set of original
strings}

\item{s_2}{A vector or matrix of characters representing a second set of
original strings for comparison. If missing, a lower
triangular matrix comparing s_1 to itself will be returned.}

\item{method}{A character input identifying the string similarity method to
be used by stringsim from the 'stringdist' package}

\item{...}{Other options to be passed to stringsim}
}
\value{
A matrix of similarities corresponding to the pairwise comparisons of
  elements of s_1 and s_2. Upper triangular when s_2 is missing.
}
\description{
Compute a matrix of similarity measures between strings, accounting for ambiguous
characters. In situations where all combinations or permutations of a set
of strings are to be compared, this implementation is more efficient than calling sim_func
However, it may be less memory efficient for large inputs
}
