% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ld.R
\name{sim_ld}
\alias{sim_ld}
\title{Compute levenstein edit similarity between strings, accounting for ambiguous characters and potentially aggregating similarity over multiple transformations (e.g. pinyin for phonetic similarity and four corner code for visual similarity)}
\usage{
sim_ld(s_1, s_2, aggr = "mean")
}
\arguments{
\item{s_1}{A vector or matrix of characters representing one set of original strings}

\item{s_2}{A vector or matrix of characters representing a second set of original strings (must be the same length as s_1)}

\item{aggr}{A character function name or function object indicating how similarities calculated over multiple transformations should be aggregated}
}
\value{
A vector of edit similarities corresponding to the pairwise comparisons of elements of s_1 and s_2
}
\description{
Compute levenstein edit similarity between strings, accounting for ambiguous characters and potentially aggregating similarity over multiple transformations (e.g. pinyin for phonetic similarity and four corner code for visual similarity)
}
