% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str100lib_dataset.R
\docType{data}
\name{str100lib}
\alias{str100lib}
\title{Character structure dataset after full decomposition}
\format{An object of class \code{environment} of length 20225.}
\source{
https://github.com/cjkvi/cjkvi-ids/blob/master/ids.txt
Further processed as in the ids2struct_rad script file
}
\usage{
str100lib
}
\description{
Character structure dataset after full decomposition
}
\keyword{datasets}
