% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str1lib_dataset.R
\docType{data}
\name{str1lib}
\alias{str1lib}
\title{Character structure dataset after a single round of decomposition}
\format{An object of class \code{environment} of length 20225.}
\source{
https://github.com/cjkvi/cjkvi-ids/blob/master/ids.txt
Further processed as in the ids2struct_rad script file
}
\usage{
str1lib
}
\description{
Character structure dataset after a single round of decomposition
}
\keyword{datasets}
