% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBlib_dataset.R
\docType{data}
\name{WBlib}
\alias{WBlib}
\title{Four Corner database}
\format{An environment with Chinese characters as the names.}
\source{
\href{https://github.com/erstern/98WuBi}{erstern/98WuBi}
}
\usage{
data(WBlib)
}
\description{
A lookup table for the wubi code of Chinese characters
}
\examples{
data(WBlib)

}
\keyword{datasets}
