% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yi_check.R
\name{yi_check}
\alias{yi_check}
\title{Functions to indicate the presence of characters signalling uncertainty}
\usage{
yi_check(n)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}
}
\value{
A logical vector of the same length as n
}
\description{
Functions to indicate the presence of characters signalling uncertainty
}
\details{
yi_check checks for commonly substituted Yi names
}
