% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_4corner.R
\name{ChStr2fc}
\alias{ChStr2fc}
\title{Convert Chinese strings to four corner code.}
\usage{
ChStr2fc(Chin.strs = "", sep = "_", parallel = FALSE)
}
\arguments{
\item{sep}{Character used to seperate different characters}

\item{parallel}{Whether or not use parallel calculation}

\item{Chin.str}{The string need to be converted}
}
\value{
four corner code of \code{Chin.str}.
}
\description{
Convert Chinese strings to four corner code.
}
