% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pinyin.R
\name{ChStr2py}
\alias{ChStr2py}
\title{Convert Chinese strings to pinyin.}
\usage{
ChStr2py(Chin.strs, tones = TRUE, multi = TRUE, sep = "_", ...)
}
\arguments{
\item{tones}{Whether the output should be toned (T) or toneless.}

\item{multi}{Whether the output should list multiple pinyins for characters with multiple pronunciations}

\item{sep}{Character used to seperate different characters. Must be '' or '_' for proper operation}

\item{...}{Unused}

\item{Chin.str}{The string need to be converted}
}
\value{
pinyin of \code{Chin.str}.
}
\description{
Convert Chinese strings to pinyin.
}
\examples{
ChStr2py(c("\\u6d77\\u4e0a\\u751f\\u660e\\u6708","\\u5929\\u6daf\\u5171\\u6b64\\u65f6"))
}
