% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pinyin.R
\name{ChStr2py}
\alias{ChStr2py}
\title{Convert Chinese strings to pinyin.}
\usage{
ChStr2py(Chin.strs = "", method = c("toneless", "tone"),
  multi = FALSE, sep = "_", parallel = FALSE)
}
\arguments{
\item{method}{Whether the output should be toned or toneless.}

\item{multi}{Whether the output should contain more than one pinyin}

\item{sep}{Character used to seperate different characters}

\item{parallel}{Whether or not use parallel calculation}

\item{Chin.str}{The string need to be converted}
}
\value{
pinyin of \code{Chin.str}.
}
\description{
Convert Chinese strings to pinyin.
}
