% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_radicals.R
\encoding{UTF-8}
\name{ChStr2rad}
\alias{ChStr2rad}
\title{Extract radical decomposition of Chinese Character strings.}
\usage{
ChStr2rad(Chin.strs, sep = "", structure = FALSE, ....)
}
\arguments{
\item{Chin.strs}{A vector of Chinese character strings}

\item{sep}{Separating character to be placed between results for individual
Hanzi. Must be '_' or '' for proper functionality}

\item{structure}{whether to append ideographic structure characters at the end of the string}

\item{...}{unused}
}
\value{
A vector of radical decomposition strings from the original input
}
\description{
Extract radical decomposition of Chinese Character strings.
}
\examples{
ChStr2rad('\\u51e8\\u51aa',structure=F)
ChStr2rad('\\u51e8\\u51aa',structure=T)
}
