% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chstr2struct.R
\name{Chstr2struct}
\alias{Chstr2struct}
\title{Extract structure information from Chinese Characters strings}
\usage{
Chstr2struct(n, times = 1)
}
\arguments{
\item{n}{A vector of Chinese character strings}

\item{times}{Maximum number of serial decompositions to use}
}
\value{
A vector of unicode ideograph structure strings representing the
 structure of the input characters, processed through at most \code{times}
 rounds of decomposition
}
\description{
Extract structure information from Chinese Characters strings
}
\examples{
Chstr2struct('凨冪',1)
Chstr2struct('凨冪',2)
}
