% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wubi.R
\encoding{UTF-8}
\name{ChStr2wb}
\alias{ChStr2wb}
\title{Convert Chinese strings to wubi code (based on radicals).}
\usage{
ChStr2wb(Chin.strs, sep = "_", ...)
}
\arguments{
\item{Chin.strs}{The string need to be converted}

\item{sep}{Character used to seperate different characters, must be '_' or '' for proper functioning}

\item{...}{unused}
}
\value{
wubi code of \code{Chin.str}.
}
\description{
Convert Chinese strings to wubi code (based on radicals).
}
\examples{
\dontrun{ChStr2wb(c("海上生明月","天涯共此时"))}
}
