% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wubi.R
\name{ChStr2wb}
\alias{ChStr2wb}
\title{Convert Chinese strings to wubi code (based on radicals).}
\usage{
ChStr2wb(Chin.strs = "", sep = "_", parallel = FALSE)
}
\arguments{
\item{sep}{Character used to seperate different characters}

\item{parallel}{Whether or not use parallel calculation}

\item{Chin.str}{The string need to be converted}
}
\value{
wubi code of \code{Chin.str}.
}
\description{
Convert Chinese strings to wubi code (based on radicals).
}
\examples{
ChStr2wb(c("海上生明月1","天涯共此时"))
}
