% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannames_data.R
\docType{data}
\name{hannames}
\alias{hannames}
\title{Baidu's top 400 most common Han family names, 2013}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 1 columns.}
\usage{
hannames
}
\description{
Baidu's top 400 most common Han family names, 2013
}
\keyword{datasets}
