% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_py.R
\encoding{UTF-8}
\name{revpy}
\alias{revpy}
\title{Function to parse format used for storing character matches when original
names are written partially in pinyin}
\usage{
revpy(n, all = F, py = F)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}

\item{all}{A logical indicator. If set to TRUE, all possible names are
returned. If false, only the first character match is returned.}

\item{py}{A logical indicator. If set to TRUE, the original pinyin element is
returned instead of possible characters}
}
\value{
A list of parsed names
}
\description{
Function to parse format used for storing character matches when original
names are written partially in pinyin
}
\examples{
#original name: chen1\\u6e9c
\dontrun{revpy("{chen1,[e,h+,g,h3,h,,f
;,i4,g,f5,g(,g8,g6,f#=]}f:", all=F)}
\dontrun{revpy("{chen1,[e,h+,g,h3,h,,f
;,i4,g,f5,g(,g8,g6,f#=]}f:", all=T)}
\dontrun{revpy("{chen1,[e,h+,g,h3,h,,f
;,i4,g,f5,g(,g8,g6,f#=]}f:", py=T)}
}
