% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_py.R
\name{revpy}
\alias{revpy}
\title{Function to parse format used for storing character matches when original
names are written partially in pinyin}
\usage{
revpy(n, all = F)
}
\arguments{
\item{n}{A string or vector of strings. Should be Chinese names.}

\item{all}{A logical indicator. If set to TRUE, all possible names are
returned. If false, only the first character match is returned.}
}
\value{
A list of parsed names
}
\description{
Function to parse format used for storing character matches when original
names are written partially in pinyin
}
\examples{
#original name: chen1溜
revpy('{chen1,[嗔,諃,琛,賝,謓,抻,郴,瞋,捵,瘨,縝,綝,棽]}溜', all=F)
revpy('{chen1,[嗔,諃,琛,賝,謓,抻,郴,瞋,捵,瘨,縝,綝,棽]}溜', all=F)
}
